+++ src/App.js
@@ -1,11 +1,20 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Switch, Route } from 'react-router-dom';
 import 'materialize-css/dist/css/materialize.min.css';
 import './App.css';
 import HomePage from './pages/Home';
 import MembersPage from './pages/Members';
+import M from 'materialize-css';
 
 const App = () => {
+    // const initial = () => {
+    //     M.AutoInit();
+    // };
+
+    // useEffect(() => {
+    //     initial();
+    // }, [initial]);
+
     return (
         <div>
             <Switch>
